Add option for DSOP Arcade controls
There is an option to change the controller group in
"Assets>Scripts>ScriptableObjects>ControlSettings" called Controller
Group. This uses the Enum Controller to append a tens place to the
controller numbers so one group is 0 and that is controls labeled as 1-4
there is now DSOP Arcade which will be controls labelled 11-14. If a
new set of controls were added then you would label them 21-24 and add
the title to the Controller Enum with the value = 20.